requirements.txt
textual>=0.47.0
GitPython>=3.1.40
